analytics_duckdb/requirements.txt
duckdb>=1.1.3
aiohttp>=3.9.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
